fix(composability): split DnD axes and close install/runtime seams
Ship row/column (and virtualized) DnD as separate structure modules and
registry packages, move min column size into constants, default sorting
off until detected, and bring DnD bodies to context-menu/resize parity.
Remove the contributing audit page in favor of the fixes.
