+++ astro.config.mts
@@ -315,10 +315,6 @@ export default defineConfig({
               label: "Documentation Guidelines",
               slug: "contributing/documentation-guidelines",
             },
-            {
-              label: "Composability Audit",
-              slug: "contributing/composability-audit",
-            },
           ],
         },
       ],
