+++ registry.json
@@ -739,7 +739,7 @@
       "name": "data-table-row-dnd",
       "type": "registry:component",
       "title": "DataTable Row DnD",
-      "description": "Drag-and-drop row reordering with composable primitives. Includes DnD provider, draggable row, drag handle, and DnD-aware table body.",
+      "description": "Drag-and-drop row reordering with composable primitives. Includes DnD provider, draggable row, drag handle, and row DnD table body. Does not ship column DnD.",
       "dependencies": [
         "@dnd-kit/core",
         "@dnd-kit/modifiers",
@@ -758,19 +758,14 @@
           "target": "components/niko-table/filters/table-row-dnd.tsx"
         },
         {
-          "path": "src/components/niko-table/filters/table-column-dnd.tsx",
-          "type": "registry:component",
-          "target": "components/niko-table/filters/table-column-dnd.tsx"
-        },
-        {
           "path": "src/components/niko-table/components/data-table-row-dnd.tsx",
           "type": "registry:component",
           "target": "components/niko-table/components/data-table-row-dnd.tsx"
         },
         {
-          "path": "src/components/niko-table/core/data-table-dnd-structure.tsx",
+          "path": "src/components/niko-table/core/data-table-row-dnd-structure.tsx",
           "type": "registry:component",
-          "target": "components/niko-table/core/data-table-dnd-structure.tsx"
+          "target": "components/niko-table/core/data-table-row-dnd-structure.tsx"
         }
       ]
     },
@@ -778,7 +773,7 @@
       "name": "data-table-column-dnd",
       "type": "registry:component",
       "title": "DataTable Column DnD",
-      "description": "Drag-and-drop column reordering with composable primitives. Includes DnD provider, draggable header, drag-along cell, and DnD-aware header/body.",
+      "description": "Drag-and-drop column reordering with composable primitives. Includes DnD provider, draggable header, drag-along cell, and column DnD header/body. Does not ship row DnD.",
       "dependencies": [
         "@dnd-kit/core",
         "@dnd-kit/modifiers",
@@ -793,19 +788,58 @@
           "target": "components/niko-table/filters/table-column-dnd.tsx"
         },
         {
-          "path": "src/components/niko-table/filters/table-row-dnd.tsx",
-          "type": "registry:component",
-          "target": "components/niko-table/filters/table-row-dnd.tsx"
-        },
-        {
           "path": "src/components/niko-table/components/data-table-column-dnd.tsx",
           "type": "registry:component",
           "target": "components/niko-table/components/data-table-column-dnd.tsx"
         },
         {
-          "path": "src/components/niko-table/core/data-table-dnd-structure.tsx",
+          "path": "src/components/niko-table/core/data-table-column-dnd-structure.tsx",
           "type": "registry:component",
-          "target": "components/niko-table/core/data-table-dnd-structure.tsx"
+          "target": "components/niko-table/core/data-table-column-dnd-structure.tsx"
+        }
+      ]
+    },
+    {
+      "name": "data-table-virtualized-row-dnd",
+      "type": "registry:component",
+      "title": "DataTable Virtualized Row DnD",
+      "description": "Virtualized row drag-and-drop body (`DataTableVirtualizedDndBody`). Requires data-table-virtualized + data-table-row-dnd. Does not ship column DnD.",
+      "dependencies": [
+        "@tanstack/react-virtual",
+        "@dnd-kit/sortable",
+        "@dnd-kit/utilities"
+      ],
+      "registryDependencies": [
+        "https://niko-table.com/r/data-table-virtualized.json",
+        "https://niko-table.com/r/data-table-row-dnd.json"
+      ],
+      "files": [
+        {
+          "path": "src/components/niko-table/core/data-table-virtualized-row-dnd-structure.tsx",
+          "type": "registry:component",
+          "target": "components/niko-table/core/data-table-virtualized-row-dnd-structure.tsx"
+        }
+      ]
+    },
+    {
+      "name": "data-table-virtualized-column-dnd",
+      "type": "registry:component",
+      "title": "DataTable Virtualized Column DnD",
+      "description": "Virtualized column drag-and-drop header/body (`DataTableVirtualizedDndHeader`, `DataTableVirtualizedDndColumnBody`). Requires data-table-virtualized + data-table-column-dnd. Does not ship row DnD.",
+      "dependencies": [
+        "@tanstack/react-virtual",
+        "@dnd-kit/sortable",
+        "@dnd-kit/utilities"
+      ],
+      "registryDependencies": [
+        "https://niko-table.com/r/data-table-virtualized.json",
+        "https://niko-table.com/r/data-table-column-dnd.json"
+      ],
+      "files": [
+        {
+          "path": "src/components/niko-table/core/data-table-virtualized-column-dnd-structure.tsx",
+          "type": "registry:component",
+          "target": "components/niko-table/core/data-table-virtualized-column-dnd-structure.tsx"
         }
       ]
     },
