+++ src/components/niko-table/lib/column-resize-handle.tsx
@@ -19,14 +19,15 @@
  */
 import { type Header } from "@tanstack/react-table"
 import { cn } from "@/lib/utils"
+import { DEFAULT_MAX_COLUMN_SIZE, DEFAULT_MIN_COLUMN_SIZE } from "./constants"
 
 /** Room for cell chrome the text measurement doesn't cover. */
 const AUTOSIZE_CELL_PADDING = 24 // horizontal padding + buffer
 const AUTOSIZE_HEADER_CHROME = 52 // sort/menu trigger + resize grip
 
-/** Column-width bounds used when a column omits `minSize` / `maxSize`. */
-export const DEFAULT_MIN_COLUMN_SIZE = 40
-export const DEFAULT_MAX_COLUMN_SIZE = 1000
+/** Re-export for consumers that imported sizes from this module. */
+export { DEFAULT_MAX_COLUMN_SIZE, DEFAULT_MIN_COLUMN_SIZE }
+
 /** Keyboard resize step (px); the larger step applies while Shift is held. */
 const KEYBOARD_RESIZE_STEP = 8
 const KEYBOARD_RESIZE_STEP_LARGE = 40
