+++ src/content/docs/contributing/index.mdx
@@ -13,8 +13,4 @@ import { LinkCard, CardGrid } from "@astrojs/starlight/components"
     title="Documentation Guidelines"
     href="/contributing/documentation-guidelines/"
   />
-  <LinkCard
-    title="Composability Audit"
-    href="/contributing/composability-audit/"
-  />
 </CardGrid>
