+++ src/content/docs/examples/column-dnd-table.mdx
@@ -31,7 +31,7 @@ Column DnD lets users reorder table columns by dragging headers. Built with `@dn
 
 Install the DataTable core and column DnD add-ons. Add `data-table-view-dnd-menu` and `data-table-search-filter` for the toolbar shown in the example:
 
-<CliCommandCode action="run" command="shadcn@latest add @niko-table/data-table @niko-table/data-table-column-dnd @niko-table/data-table-view-dnd-menu @niko-table/data-table-search-filter" />
+<CliCommandCode action="run" command="shadcn@latest add @niko-table/data-table @niko-table/data-table-column-dnd @niko-table/data-table-column-resize @niko-table/data-table-view-dnd-menu @niko-table/data-table-search-filter" />
 
 > First time using `@niko-table`? See the [Installation Guide](/getting-started/installation#configure-the-niko-table-registry) to set up the registry.
