+++ src/content/docs/examples/row-dnd-table.mdx
@@ -31,7 +31,7 @@ Row DnD lets users reorder table rows by dragging. Built with `@dnd-kit` and com
 
 Install the DataTable core and row DnD add-on:
 
-<CliCommandCode action="run" command="shadcn@latest add @niko-table/data-table @niko-table/data-table-row-dnd @niko-table/data-table-column-resize" />
+<CliCommandCode action="run" command="shadcn@latest add @niko-table/data-table @niko-table/data-table-row-dnd @niko-table/data-table-column-resize @niko-table/data-table-virtualized @niko-table/data-table-virtualized-row-dnd" />
 
 > First time using `@niko-table`? See the [Installation Guide](/getting-started/installation#configure-the-niko-table-registry) to set up the registry.
